[refactor] 스타일 변경 시 가장 상위 feature도 같이 입력받도록 함 | action Type위치 변경
- 스타일 변경 할 때 가장 상위 feature도 payload값으로 같이 입력받음(구별 위함)
- action Type의 위치를 다른 type들이 같이 있는 파일 안으로 이동
